Feature/recommend ranking 热门推荐体系，包括Sniffer、Recommender、Service；推荐队列Service

+++ src/main/java/com/chenlinghong/graduation/repository/domain/RecommendRankingGoods.java
@@ -30,5 +30,5 @@ public class RecommendRankingGoods extends BaseDomain {
     /**
      * 排行榜
      */
-    private Integer ranking;
+    private Double ranking;
 }
